+++ app/src/main/resources/io/xpipe/app/resources/lang/dscreation_en.properties
@@ -23,10 +23,14 @@ anyStream=Any Stream Type
 noMatchingStoreFound=No suitable saved store was found
 addStore=Add Store
 anyStreamDescription=Or choose specific type
+restart=Restart XPipe
+restartDescription=A restart can often be a quick fix
 reportIssue=Report Issue
 reportIssueDescription=Open the integrated issue reporter
 usefulActions=Useful actions
 stored=Saved
+troubleshootingOptions=Troubleshooting options
+troubleshoot=Troubleshoot
 other=Other
 remote=Remote File
 addShellStore=Add Shell ...
