+++ app/src/main/resources/io/xpipe/app/resources/lang/preferences_en.properties
@@ -44,6 +44,7 @@ startGui=Start GUI
 startInTray=Start in tray
 startInBackground=Start in background
 clearCaches=Clear caches ...
+clearCachesDescription=Delete all cache data
 ok=OK
 apply=Apply
 cancel=Cancel
